Airline Project - 0.2
Updated the format of the code to have more complex questions. The program now asks the user if they wish to fly in economy or business and them generates a random price based on this. Will be looking to add data validation to this project next, afterwards i will be looking to refactor the code into a neater format.